Eliminating all references to JFreeChart
